feat: change subtotal, delivery_fee, and total_amount fields to decimal in checkout and order models

+++ backend/apps/checkout/models.py
@@ -55,11 +55,15 @@ class CheckoutTransaction(models.Model):
         blank=True,
     )
 
-    subtotal = models.IntegerField(
+    subtotal = models.DecimalField(
+        max_digits=12,
+        decimal_places=2,
         default=0,
     )
 
-    delivery_fee = models.IntegerField(
+    delivery_fee = models.DecimalField(
+        max_digits=12,
+        decimal_places=2,
         default=0,
     )
